+++ OrchestratorV2/interfaces/task/parameters.ts
@@ -1,6 +1,7 @@
 import { IFilters } from "./filters";
 import { IReleaseVariable } from "../common/releasevariable";
 import { ReleaseType } from "../common/releasetype";
+import { ISettings } from "../common/settings";
 
 export interface IParameters {
 
@@ -11,5 +12,6 @@ export interface IParameters {
     stages: string[];
     variables: IReleaseVariable[];
     filters: IFilters;
+    settings: ISettings;
 
 }