+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## 1.3.0
+## 2.0.0
 
+- Major internal refactoring (TBU)
+- Update task to use name parameters ([4](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/4), [43](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/43))
 - Add latest release stage status filter support ([26](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/26))
 - Add release variables support ([24](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/24))
 - Make approval retries configurable ([23](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/23))
-- Major internal refactoring (TBU)
 
 ## 1.2.9